refactor: 重構專案結構並整合 Copilot 設定
- 將核心類別移至 src/core 目錄（GoogleSheetLanguagesModel 與 LanguagesModel）
- 更新所有相關檔案的 import 路徑
- 新增 git submodule 引入 vscode-copilot-prompts 通用提示
- 設定 VSCode Copilot 提示檔案位置與 commit message 生成指示
- 移除舊有的 copilot-commit-message-instructions.md 檔案
- 更新專案相依套件至最新版本
