+++ .vscode/settings.json
@@ -1,0 +1,11 @@
+{
+  "chat.promptFilesLocations": {
+    ".github/prompts": true,                          // 專案特定的 prompts
+    ".github/copilot-common/prompts": true            // 通用的 prompts
+  },
+  "github.copilot.chat.commitMessageGeneration.instructions": [
+    {
+      "file": ".github/copilot-common/instructions/commit-message-generation.instructions.md"
+    }
+  ]
+}
