+++ package.json
@@ -23,7 +23,8 @@
   "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "scripts": {
-    "test": "vitest",
+    "test": "vitest run",
+    "test:watch": "vitest",
     "dev": "tsdown --watch",
     "build": "tsdown",
     "i18n:push": "tsx example/push.ts",
@@ -32,16 +33,18 @@
   },
   "dependencies": {
     "googleapis": "^150.0.1",
-    "lodash-es": "^4.17.21"
+    "lodash-es": "^4.17.21",
+    "yargs": "^18.0.0"
   },
   "devDependencies": {
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "^20.17.10",
-    "dotenv": "^16.4.7",
-    "tsdown": "^0.13.0",
-    "tsx": "^4.19.2",
-    "typescript": "^5.7.2",
-    "vitest": "^2.1.8"
+    "@types/node": "^24.9.1",
+    "@types/yargs": "^17.0.33",
+    "dotenv": "^17.2.3",
+    "tsdown": "^0.15.9",
+    "tsx": "^4.20.6",
+    "typescript": "^5.9.3",
+    "vitest": "^3.2.4"
   },
   "peerDependencies": {
     "googleapis": "^150.0.1"
